+++ src/inotify.rs
@@ -404,9 +404,9 @@ impl EventLoop {
     }
 
     fn add_watch(&mut self, path: PathBuf, is_recursive: bool, mut watch_self: bool) -> Result<()> {
-        let metadata = metadata(&path).map_err(Error::io)?;
-
-        if !metadata.is_dir() || !is_recursive {
+        // If the watch is not recursive, or if we determine (by stat'ing the path to get its
+        // metadata) that the watched path is not a directory, add a single path watch.
+        if !is_recursive || !metadata(&path).map_err(Error::io)?.is_dir() {
             return self.add_single_watch(path, false, true);
         }